make territory revenue idependent job
